Add pitch detection with crepe

+++ src/separator.py
@@ -3,6 +3,8 @@ import os
 from pathlib import Path
 import ffmpeg
 
+REUSE_OK = True
+
 
 def separate(path: str, model: str = "htdemucs_ft", output: str = "out") -> Path:
     result = Path(output) / model / os.path.splitext(Path(path).name)[0]
@@ -18,14 +20,19 @@ def separate(path: str, model: str = "htdemucs_ft", output: str = "out") -> Path
     return result
 
 
-def rename_stem(path: str | Path, stem: str = "no_vocals.mp3") -> str:
+def convert(
+    path: str | Path, stem: str = "no_vocals.mp3", extension: str = "ogg"
+) -> str:
     path = Path(path)
     if stem == "no_vocals.mp3":
         stem_type = "INSTRUMENTAL"
     else:
         stem_type = "VOCALS"
-    result = f"{path.name} [{stem_type}].ogg"
+    result = f"{path.name} [{stem_type}].{extension}"
     if os.path.exists(path / result):
+        if REUSE_OK:
+            print("Reusing previous result")
+            return result
         print("Deleting old result file")
         os.remove(path / result)
     worker = ffmpeg.FFmpeg().input(path / stem).output(path / result)
